Moving doubles to floats

+++ Det/FTDet/src/Lib/DeFTLayer.cpp
@@ -53,7 +53,7 @@ StatusCode DeFTLayer::initialize() {
   const Gaudi::XYZPoint g3 = geometry() -> toGlobal( Gaudi::XYZPoint(0.,1.,0.) );
   m_plane = Gaudi::Plane3D(g1,g2,g3 );
   m_dzdy = (g3.z() - g1.z()) / (g3.y() - g1.y());
-  m_stereoAngle = (double)param<double>("stereoAngle");
+  m_stereoAngle = (float)param<double>("stereoAngle");
   m_dxdy = -tan( m_stereoAngle );
 
   return StatusCode::SUCCESS;
